requirements.txt
click==8.1.7
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
ollama==0.1.0
pydantic==2.5.0
